+++ src/Utils.h
@@ -1,7 +1,7 @@
 #pragma once
 
 void IntToString(char *buffer, size_t bufferSize, int value);
-int Random(int max);
+uint16_t Random(uint16_t max);
 void SetRandomSeed(long newSeed);
 
 unsigned int GetUnixTime(PblTm *current_time);
